+++ src/functions/interpret.js
@@ -9,6 +9,8 @@ const { _addToHist } = require("../commands/history");
 
 // Get classes
 const Errors = require("../classes/Errors");
+const Verbose = require("../classes/Verbose");
+const InfoMessages = require("../classes/InfoMessages");
 
 /**
  * Split the commands passed into BubbleOS, and then remove the actual command name (the first word)
@@ -30,11 +32,14 @@ const getKeyByValue = (object, value) => Object.keys(object).find((key) => objec
  * Interpret all available BubbleOS commands.
  *
  * @param {string} command The command that was requested to be interpreted by the user.
+ * @param {boolean} storeInHistory Whether or not to store the command in history. Defaults to true.
  */
-const _intCmds = async (command) => {
+const _intCmds = async (command, storeInHistory = true) => {
   try {
     // If the command is empty or not
+    Verbose.custom("Checking if command entered is empty...");
     const isEmpty = command.length === 0;
+    Verbose.custom("Checking entered command...");
     const enteredCmd = command.split(" ")[0];
 
     // The command is currently unrecognized
@@ -44,9 +49,11 @@ const _intCmds = async (command) => {
     for (let [key, value] of Object.entries(commands)) {
       // If the command starts with the current command
       if (enteredCmd === key) {
-        // Make the command recognized
+        Verbose.custom("Command has been recognized.");
         recognized = true;
+
         // If the command is 'bub', it requires the '_intCmds' function, so call/pass it separately
+        Verbose.custom("Executing command...");
         if (key === "bub") {
           await value(_intCmds, ..._multiParam(command));
         } else {
@@ -57,17 +64,19 @@ const _intCmds = async (command) => {
 
     // If the command is not recognized and isn't empty
     if (!recognized && !isEmpty) {
+      Verbose.custom(
+        `Command '${enteredCmd}' was detected to be unrecognized, show respective error...`
+      );
       Errors.unrecognizedCommand(enteredCmd);
 
       for (const alias of Object.values(aliases)) {
         for (const cmd of alias) {
           if (cmd === enteredCmd) {
-            console.log(
-              chalk.yellow(
-                `There is no command called ${chalk.italic(enteredCmd)}. Did you mean ${chalk.bold(
-                  getKeyByValue(aliases, alias)
-                )}?\n`
-              )
+            Verbose.custom("Alias detected for entered command, show tip to user...");
+            InfoMessages.info(
+              `There is no command called ${chalk.italic(enteredCmd)}. Did you mean ${chalk.bold(
+                getKeyByValue(aliases, alias)
+              )}?`
             );
 
             break;
@@ -76,8 +85,12 @@ const _intCmds = async (command) => {
       }
     }
 
-    if (!isEmpty && command !== "history -c") _addToHist(command);
+    if (!isEmpty && storeInHistory && command !== "history -c") {
+      Verbose.custom("Adding command to history...");
+      _addToHist(command);
+    }
   } catch (err) {
+    Verbose.fatalError();
     _fatalError(err);
   }
 };